% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spindler-package.R
\docType{package}
\name{spindler-package}
\alias{spindler}
\alias{spindler-package}
\title{spindler: Create and Publish Twitter Threads}
\description{
Publish Twitter threads from your R session or from a rmarkdown document.
}
\section{Setup}{

In order for spindler to post on your Twitter account you'll need to create a
Twitter App and authenticate it with rtweet. Check https://rtweet.info/articles/auth.html
for instructions.
}

\section{Functions}{

There's only two functions.
\itemize{
\item \link{thread}
\item \code{\link[=knitr_thread]{knitr_thread()}} sets up a rmarkdown document so you can publish chunks as
posts in a thread.
}
}

\author{
\strong{Maintainer}: Elio Campitelli \email{elio.campitelli@cima.fcen.uba.ar} (0000-0002-7742-9230)

}
\keyword{internal}
