% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renders.R
\name{spindler_render}
\alias{spindler_render}
\title{Render media}
\usage{
spindler_render(object, ...)
}
\arguments{
\item{object}{R object to render.}

\item{...}{Other arguments passed to the renderer function.}
}
\value{
The path to the rendered media.
}
\description{
Methods for rendering custom content.
}
