% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadr.R
\name{thread}
\alias{thread}
\alias{new}
\alias{add_post}
\alias{show}
\alias{get_url}
\alias{$}
\title{Twitter thread}
\usage{
thread$new(tag = NULL, publish = FALSE)
thread$add_post(status, media = NULL)
thread$clear()
thread$publish()
thread$browse(n = 1)
thread$get_url(n = 1)
thread$get_posts()
thread$destroy()
thread$add_whatermark()
}
\arguments{
\item{tag}{String that can be used as a chunk option inside a rmarkdown document
to populate the thread (see Examples.)}

\item{publish}{Wheter to publish the thread when finishing rendering. Only used
when rendering a rmarkdown document.}

\item{status}{Text to be tweeted.}

\item{media}{Path to an image or video that will be attached to the tweet.}

\item{n}{Number of post in thread to show in browser or get url from.}
}
\description{
Creates a twitter thread object that can be populated with posts, published and destroyed.
}
\details{
The basic workflow is to create a new thread object with \code{thread$new()} and
then populate it with \code{thread$add_post()}. Once you are happy with it, publish
it to Tweeter with \code{thread$publish()}. If you want to "unpublish" it, use
\code{thread$destroy()}. This will delete each post on Twitter, but they will
still be saved on your thread object. Use \code{thread$clear()} to delete them.

To view the list of posts currently on your thread, use \code{thread$get_posts()}.
Once published, you can get the url of each post with \code{thread$get_url()} or
open it up in a browser session with \code{thread$browse()}.

The \code{thread$add_watermark()} method adds this post:

This thread comes to you courtesy of the spindler \U1F4E6 \cr
Reproducible tweets with R and rmarkdown. \cr
#rstats \cr
https://git.io/fjzxN

If you like the package, consider adding it so more people can enjoy it :).
}
\examples{
\dontrun{
birds <- thread$new()
birds$add_post("Hey, people, I want to tell you how awesome birds are!")$
  add_post("They have feathers, and (most of them) can fly!")$
  add_post("And look how cute they ares", media = "~/Pictures/penguin1.png")

birds$publish()

# Oh, no I made a typo. Quick, delete the whole thing!
birds$destroy()

# Let's start over
birds$clear()$
  add_post("Nooo! I had an awesome thread about birds, but I messed up.")$
  add_post("So here's the jist of it: birds rock and they are better than monkeys!")$
  publish

# Look at the finished product
birds$browse()

# You can use the tag to populate a thread automatically from a chunk.
# The first figure produced by the chunk will be attached as media.
```{r, setup}
this_thread <- spindler::thread$new(tag = "tw_status", publish = TRUE)
```

```{r, tw_status = "The relationship between pressure and temperature is cool!"}
plot(pressure)
```

```{r}
this_thread$add_post("This post is a free agent, not tied to any chunk.")
```
}

}
