% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-thread.R, R/twitter.R
\name{saved_threads}
\alias{saved_threads}
\title{Load all saved threads.}
\usage{
saved_threads()

saved_threads()
}
\value{
A list with \link{thread} objects.

A list with \link{thread} objects.
}
\description{
Loads all the thread objects in a directory.

Loads all the thread objects in a directory.
}
\examples{
\dontrun{
# Create a thread, publish it and save it.
thread$new()$
   add_post("This is a test thread!")$
   publish()$
   save()

# List all saved threads
threads <- saved_threads()

# Destroy the last thread.
threasd[[length(threads)]]$destroy()
}

\dontrun{
# Create a thread, publish it and save it.
thread$new()$
   add_post("This is a test thread!")$
   publish()$
   save()

# List all saved threads
threads <- saved_threads()

# Destroy the last thread.
threasd[[length(threads)]]$destroy()
}

}
