% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spindler-package.R
\docType{package}
\name{spindler-package}
\alias{spindler}
\alias{spindler-package}
\title{spindler: Create and Publish Twitter Threads}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Publish Twitter threads from your R session or from a rmarkdown document.
}
\section{Setup}{

In order for spindler to post on your Twitter account you'll need to create a
Twitter App and authenticate it with rtweet. Check https://rtweet.info/articles/auth.html
for instructions.
}

\section{Usage}{

All you need to know is on the \link{thread} object.
}

\author{
\strong{Maintainer}: Elio Campitelli \email{elio.campitelli@cima.fcen.uba.ar} (\href{https://orcid.org/0000-0002-7742-9230}{ORCID})

}
\keyword{internal}
