% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadr.R
\docType{data}
\name{thread}
\alias{thread}
\title{Twitter thread}
\description{
Creates a twitter thread object that can be populated with posts, published and destroyed.
}
\section{Creating}{

You can create a new thread with \code{thread$new()}. No parameters necessary.
}

\section{Populating}{

Each post is added in order with \code{thread$add_post(status, media)}. Where \code{status}
is a length one character vector and media is a length one character vector with
the path to the media that will be uploaded.
}

\section{Cleaning up}{

If you made a mistake or want to start over for some reason, use \code{thread$clear()}.
}

\section{Publishing}{

When you are happy with how your thread has turned up, publish it to Twitter
with \code{thread$publish()}. The url of each tweet can be retrieved with \code{thread$get_url()}.

To view the result, use \code{thread$show()}.
}

\section{Deleting}{

To delete the thread from your timeline, use \code{thread$destroy()}. This will
delete each post on Twitter, but it will still be saved on your thread object.
}

\examples{
\dontrun{
birds <- thread$new()
birds$add_post("Hey, people, I want to tell you how awesome birds are!")$
  add_post("They have feathers, and (most of them) can fly!")$
  add_post("And look how cute they ares", media = "pictures/birds/penguin1.png")

birds$publish()

# Oh, no I made a typo. Quick, delete the whole thing!
birds$destroy()

# Lets start over
birds$clear()$
  add_post("Nooo! I had an awesome thread about birds, but I messed up.")$
  add_post("So here's the jist of it: birds rock and they are better than monkeys!")$
  publish

# Look at the finished product
birds$show()
}

}
\keyword{datasets}
